Refactor: Move RuleSetRunner into a separate file
The RuleSetRunner class moved to a new file rule_set_runner.py
